+++ packages/openapi-types/src/openapi-types.test-d.ts
@@ -1,72 +1,84 @@
 import { describe, expectTypeOf, it } from 'vitest'
 
-import type { OpenAPI, OpenAPIV2, OpenAPIV3, OpenAPIV3_1 } from './openapi-types'
+import type { OpenAPI, OpenAPIV2, OpenAPIV3, OpenAPIV3_1, OpenAPIV3_2 } from './openapi-types'
 
 describe('OpenAPI', () => {
   it('has a generic type', () => {
-    const specification: OpenAPI.Document = {
+    const document: OpenAPI.Document = {
       // anything
     }
 
-    expectTypeOf(specification).toMatchTypeOf<OpenAPI.Document>()
+    expectTypeOf(document).toEqualTypeOf<OpenAPI.Document>()
   })
 
   it('narrows it down to Swagger 2.0', () => {
-    const specification: OpenAPI.Document = {
+    const document: OpenAPI.Document = {
       swagger: '2.0',
     }
 
-    expectTypeOf(specification).toMatchTypeOf<OpenAPIV2.Document>()
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV2.Document>()
   })
 
   it('narrows it down to OpenAPI 3.0.0', () => {
-    const specification: OpenAPI.Document = {
+    const document: OpenAPI.Document = {
       openapi: '3.0.0',
     }
 
-    expectTypeOf(specification).toMatchTypeOf<OpenAPIV3.Document>()
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV3.Document>()
   })
 
   it('narrows it down to OpenAPI 3.0.4', () => {
-    const specification: OpenAPI.Document = {
+    const document: OpenAPI.Document = {
       openapi: '3.0.4',
     }
 
-    expectTypeOf(specification).toMatchTypeOf<OpenAPIV3.Document>()
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV3.Document>()
   })
 
   it('narrows it down to OpenAPI 3.1.0', () => {
-    const specification: OpenAPI.Document = {
+    const document: OpenAPI.Document = {
       openapi: '3.1.0',
     }
 
-    expectTypeOf(specification).toMatchTypeOf<OpenAPIV3_1.Document>()
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV3_1.Document>()
   })
 
   it('narrows it down to OpenAPI 3.1.1', () => {
-    const specification: OpenAPI.Document = {
+    const document: OpenAPI.Document = {
       openapi: '3.1.1',
     }
 
-    expectTypeOf(specification).toMatchTypeOf<OpenAPIV3_1.Document>()
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV3_1.Document>()
+  })
+
+  it('narrows it down to OpenAPI 3.1.2', () => {
+    const document: OpenAPI.Document = {
+      openapi: '3.1.2',
+    }
+
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV3_1.Document>()
+  })
+
+  it('narrows it down to OpenAPI 3.2.0', () => {
+    const document: OpenAPI.Document = {
+      openapi: '3.2.0',
+    }
+
+    expectTypeOf(document).toEqualTypeOf<OpenAPIV3_2.Document>()
   })
 
   it('types a custom extension', () => {
-    const specification: OpenAPI.Document<{
+    const document: OpenAPI.Document<{
       'x-custom'?: boolean
       'random-attribute'?: any
     }> = {}
 
-    expectTypeOf(specification['random-attribute']).toEqualTypeOf<any>()
-    expectTypeOf(specification['x-custom']).toEqualTypeOf<boolean | undefined>()
+    expectTypeOf(document['random-attribute']).toEqualTypeOf<any>()
+    expectTypeOf(document['x-custom']).toEqualTypeOf<boolean | undefined>()
   })
 
   it('has a HttpMethod type', () => {
-    const validMethod: OpenAPI.HttpMethod = 'GET'
-    const anotherValidMethod: Lowercase<OpenAPI.HttpMethod> = 'get'
-
-    expectTypeOf(validMethod).toMatchTypeOf<OpenAPI.HttpMethod>()
-    expectTypeOf(anotherValidMethod).toMatchTypeOf<Lowercase<OpenAPI.HttpMethod>>()
+    expectTypeOf('get' as OpenAPI.HttpMethod).toEqualTypeOf<OpenAPI.HttpMethod>()
 
     // @ts-expect-error name is a string
     assertType('NOT_A_METHOD' as OpenAPI.HttpMethod)