+++ integrations/hono/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/hono-api-reference
 
+## 0.9.22
+
+### Patch Changes
+
+- @scalar/core@0.3.20
+
 ## 0.9.21
 
 ### Patch Changes